Create new utils.ts.
This moves the Initial type into utils.ts.

Moves creating a list of initials into an rxjs pipe in utils.ts.

Moves doing a case insensitive prefix search into being an rxjs pipe
in utils.ts
