+++ frontend/src/app/_helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './auth.guard';
 export * from './user-page-redirect.guard';
+export * from './utils';
